feat: 配置国内Gradle镜像

+++ app/src/main/java/com/silas/omaster/util/PerformanceHelper.kt
@@ -72,7 +72,7 @@ object PerformanceHelper {
     fun getAppMemoryUsageMB(): Long {
         val memoryInfo = Debug.MemoryInfo()
         Debug.getMemoryInfo(memoryInfo)
-        return memoryInfo.getTotalPss() / 1024
+        return memoryInfo.getTotalPss().toLong() / 1024L
     }
 
     /**
